requirements.txt
gitpython
setuptools
six
requests
python-dateutil
cached-property